Have Login and Registration working as expected... need to do API's for Car Pool Information and MyProfile.
Having slight issues with time due to doing this and working at the same
time... Nice little test though hehe

+++ src/CoMute.DAL/usp_User_Register_Result.cs
@@ -13,6 +13,6 @@ namespace CoMute.DAL
     
     public partial class usp_User_Register_Result
     {
-        public Nullable<decimal> UserID { get; set; }
+        public Nullable<int> UserID { get; set; }
     }
 }
